Add interactive functions
