+++ src/index.html
@@ -36,6 +36,7 @@
                 </div>
             </form>
         </li>
+        
         <li class="clear-completed">
             <button class="btn-clear">Clear all completed</button>
         </li>
